rework init sentry
