+++ src/run-replicate-incrementally.js
@@ -1,14 +1,15 @@
 const logger = require('./logger');
 const runner = require('./replicate-incrementally');
 
-require('./sentry-config');
 const Sentry = require('@sentry/node');
+const initSentry = require('./sentry-init');
 
 const extractConfigurationFromEnvironment = require ('./extract-configuration-from-environment');
 const configuration = extractConfigurationFromEnvironment();
 
 async function main() {
   try {
+    initSentry(configuration);  
     await runner.run(configuration);
   } catch (error) {
     logger.error(error);
